requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiohttp>=3.8.0
jinja2>=3.0.0
python-owasp-zap-v2.4>=0.0.20
zaproxy>=0.4.0
httpx>=0.27.0
anthropic>=0.19.3
PyQt5>=5.15
websockets>=12.0
torch>=2.2.1
sentencepiece>=0.2.0
sse-starlette>=1.6.5
nltk>=3.6.3
pytest>=6.2.5
pytest-asyncio>=0.16.0
mcp>=1.6.0